Make sdk wallet JWT login available at all pages closes #109
